+++ sdk/attestation/azure-security-attestation/inc/azure/attestation/attestation_client.hpp
@@ -115,9 +115,47 @@ namespace Azure { namespace Security { namespace Attestation {
 
   class AttestationClient final {
     // Allow client factory to access private methods in the AttestationClient object.
-    friend class AttestationClientFactory;
+    friend class AttestationClientCreator;
 
   public:
+    class AttestationClientCreator {
+      std::string m_endpoint;
+      AttestationClientOptions m_options;
+      Azure::Core::Context m_context;
+      std::shared_ptr<Core::Credentials::TokenCredential const> m_credential;
+
+    public:
+      AttestationClientCreator(
+          std::string const& endpoint,
+          std::shared_ptr<Core::Credentials::TokenCredential const> credential,
+          AttestationClientOptions const& options,
+          Azure::Core::Context const& context)
+          : m_endpoint(endpoint), m_options(options), m_credential(credential), m_context(context)
+      {
+      }
+
+      operator AttestationClient()
+      {
+        return AttestationClient::CreateConcrete(m_endpoint, m_credential, m_options, m_context);
+      }
+      operator std::unique_ptr<AttestationClient>()
+      {
+        return AttestationClient::CreatePointer(m_endpoint, m_credential, m_options, m_context);
+      }
+      operator std::unique_ptr<AttestationClient const>()
+      {
+        return AttestationClient::CreatePointer(m_endpoint, m_credential, m_options, m_context);
+      }
+      operator std::shared_ptr<AttestationClient const>()
+      {
+        return AttestationClient::CreatePointer(m_endpoint, m_credential, m_options, m_context);
+      }
+      operator std::shared_ptr<AttestationClient>()
+      {
+        return AttestationClient::CreatePointer(m_endpoint, m_credential, m_options, m_context);
+      }
+    };
+
     /** @brief Construct a new Attestation Client object
      *
      * @details Constructs a new attestation client. Follows the
@@ -128,12 +166,12 @@ namespace Azure { namespace Security { namespace Attestation {
      * @param credential The authentication method to use (required for TPM attestation). If the
      * credential parameter is not supplied, the connection will be unauthenticated.
      * @param options The options to customize the client behavior.
-     * @return std::unique_ptr<AttestationClient> The newly created client.
+     * @return The newly created client.
      */
-    static AttestationClient Create(
+    static AttestationClientCreator Create(
         std::string const& endpoint,
         std::shared_ptr<Core::Credentials::TokenCredential const> credential,
-        AttestationClientOptions options = AttestationClientOptions{},
+        AttestationClientOptions const& options = AttestationClientOptions{},
         Azure::Core::Context const& constext = Azure::Core::Context{});
 
     /** @brief Construct a new anonymous Attestation Client object
@@ -144,12 +182,12 @@ namespace Azure { namespace Security { namespace Attestation {
      *
      * @param endpoint The URL address where the client will send the requests to.
      * @param options The options to customize the client behavior.
-     * @return std::unique_ptr<AttestationClient> The newly created attestation client.
+     * @return The newly created attestation client.
      *
      * @note TPM attestation requires an authenticated attestation client.
      *
      */
-    static AttestationClient Create(
+    static AttestationClientCreator Create(
         std::string const& endpoint,
         AttestationClientOptions options = AttestationClientOptions{},
         Azure::Core::Context const& constext = Azure::Core::Context{});
@@ -263,6 +301,41 @@ namespace Azure { namespace Security { namespace Attestation {
     std::shared_ptr<Azure::Core::Http::_internal::HttpPipeline> m_pipeline;
     AttestationTokenValidationOptions m_tokenValidationOptions;
     std::vector<Models::AttestationSigner> m_attestationSigners;
+
+    /** @brief Construct a new Attestation Client object
+     *
+     * @details Constructs a new attestation client. Follows the
+     * factory pattern in [C++ Core Guidelines
+     * C.50](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#c50-use-a-factory-function-if-you-need-virtual-behavior-during-initialization)
+     *
+     * @param endpoint The URL address where the client will send the requests to.
+     * @param credential The authentication method to use (required for TPM attestation). If the
+     * credential parameter is not supplied, the connection will be unauthenticated.
+     * @param options The options to customize the client behavior.
+     * @return std::unique_ptr<AttestationClient> The newly created client.
+     */
+    static AttestationClient CreateConcrete(
+        std::string const& endpoint,
+        std::shared_ptr<Core::Credentials::TokenCredential const> credential,
+        AttestationClientOptions const& options = AttestationClientOptions{},
+        Azure::Core::Context const& constext = Azure::Core::Context{});
+    /** @brief Construct a new Attestation Client object
+     *
+     * @details Constructs a new attestation client. Follows the
+     * factory pattern in [C++ Core Guidelines
+     * C.50](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#c50-use-a-factory-function-if-you-need-virtual-behavior-during-initialization)
+     *
+     * @param endpoint The URL address where the client will send the requests to.
+     * @param credential The authentication method to use (required for TPM attestation). If the
+     * credential parameter is not supplied, the connection will be unauthenticated.
+     * @param options The options to customize the client behavior.
+     * @return std::unique_ptr<AttestationClient> The newly created client.
+     */
+    static std::unique_ptr<AttestationClient> CreatePointer(
+        std::string const& endpoint,
+        std::shared_ptr<Core::Credentials::TokenCredential const> credential,
+        AttestationClientOptions const& options = AttestationClientOptions{},
+        Azure::Core::Context const& constext = Azure::Core::Context{});
 
     /** @brief Construct a new Attestation Client object
      *