+++ sdk/attestation/azure-security-attestation/src/attestation_client.cpp
@@ -216,16 +216,13 @@ void AttestationClient::RetrieveResponseValidationCollateral(Azure::Core::Contex
  * @param credential The authentication method to use (required for TPM attestation).
  * @param options The options to customize the client behavior.
  */
-AttestationClient AttestationClient::Create(
+Azure::Security::Attestation::AttestationClient::AttestationClientCreator AttestationClient::Create(
     std::string const& endpoint,
     std::shared_ptr<Core::Credentials::TokenCredential const> credential,
-    AttestationClientOptions options,
+    AttestationClientOptions const& options,
     Azure::Core::Context const& context)
 {
-  AttestationClient returnValue(endpoint, credential, options);
-  returnValue.RetrieveResponseValidationCollateral(context);
-  // Release the client pointer from the unique pointer to let the parent manage it.
-  return returnValue;
+  return AttestationClientCreator{endpoint, credential, options, context};
 }
 
 /** @brief Construct a new anonymous Attestation Client object
@@ -235,10 +232,48 @@ AttestationClient AttestationClient::Create(
  *
  * @note TPM attestation requires an authenticated attestation client.
  */
-AttestationClient AttestationClient::Create(
+Azure::Security::Attestation::AttestationClient::AttestationClientCreator AttestationClient::Create(
     std::string const& endpoint,
     AttestationClientOptions options,
     Azure::Core::Context const& context)
 {
   return Create(endpoint, nullptr, options, context);
+}
+
+
+/** @brief Construct a new Attestation Client object
+ *
+ * @param endpoint The URL address where the client will send the requests to.
+ * @param credential The authentication method to use (required for TPM attestation).
+ * @param options The options to customize the client behavior.
+ */
+AttestationClient AttestationClient::CreateConcrete(
+    std::string const& endpoint,
+    std::shared_ptr<Core::Credentials::TokenCredential const> credential,
+    AttestationClientOptions const& options,
+    Azure::Core::Context const& context)
+{
+  AttestationClient returnValue(endpoint, credential, options);
+  returnValue.RetrieveResponseValidationCollateral(context);
+  // Release the client pointer from the unique pointer to let the parent manage it.
+  return returnValue;
+}
+
+
+/** @brief Construct a new Attestation Client object
+ *
+ * @param endpoint The URL address where the client will send the requests to.
+ * @param credential The authentication method to use (required for TPM attestation).
+ * @param options The options to customize the client behavior.
+ */
+std::unique_ptr<AttestationClient> AttestationClient::CreatePointer(
+    std::string const& endpoint,
+    std::shared_ptr<Core::Credentials::TokenCredential const> credential,
+    AttestationClientOptions const& options,
+    Azure::Core::Context const& context)
+{
+  std::unique_ptr<AttestationClient> returnValue(new AttestationClient(endpoint, credential, options));
+  returnValue->RetrieveResponseValidationCollateral(context);
+  // Release the client pointer from the unique pointer to let the parent manage it.
+  return returnValue;
 }