+++ sdk/attestation/azure-security-attestation/test/ut/attestation_test.cpp
@@ -68,14 +68,14 @@ namespace Azure { namespace Security { namespace Attestation { namespace Test {
       return returnValue;
     }
 
-    AttestationClient CreateClient()
+    std::unique_ptr<AttestationClient> CreateClient()
     {
       // `InitTestClient` takes care of setting up Record&Playback.
       auto options = InitClientOptions<Azure::Security::Attestation::AttestationClientOptions>();
       options.TokenValidationOptions = GetTokenValidationOptions();
       return AttestationClient::Create(m_endpoint, options);
     }
-    AttestationClient CreateAuthenticatedClient()
+    std::unique_ptr<AttestationClient> CreateAuthenticatedClient()
     {
       // `InitClientOptions` takes care of setting up Record&Playback.
       AttestationClientOptions options = InitClientOptions<AttestationClientOptions>();
@@ -138,13 +138,13 @@ namespace Azure { namespace Security { namespace Attestation { namespace Test {
     if (type == AttestationType::OpenEnclave)
     {
       auto report = AttestationCollateral::OpenEnclaveReport();
-      auto attestResponse = client.AttestOpenEnclave(report);
+      auto attestResponse = client->AttestOpenEnclave(report);
       ValidateAttestResponse(attestResponse);
     }
     else if (type == AttestationType::SgxEnclave)
     {
       auto quote = AttestationCollateral::SgxQuote();
-      auto attestResponse = client.AttestSgxEnclave(quote);
+      auto attestResponse = client->AttestSgxEnclave(quote);
       ValidateAttestResponse(attestResponse);
     }
   }
@@ -162,7 +162,7 @@ namespace Azure { namespace Security { namespace Attestation { namespace Test {
       AttestOpenEnclaveOptions options;
       options.RunTimeData = data;
       auto report = AttestationCollateral::OpenEnclaveReport();
-      auto attestResponse = client.AttestOpenEnclave(report, options);
+      auto attestResponse = client->AttestOpenEnclave(report, options);
       ValidateAttestResponse(attestResponse, data);
     }
     else if (type == AttestationType::SgxEnclave)
@@ -170,7 +170,7 @@ namespace Azure { namespace Security { namespace Attestation { namespace Test {
       AttestSgxEnclaveOptions options;
       options.RunTimeData = data;
       auto quote = AttestationCollateral::SgxQuote();
-      auto attestResponse = client.AttestSgxEnclave(quote, options);
+      auto attestResponse = client->AttestSgxEnclave(quote, options);
       ValidateAttestResponse(attestResponse, data);
     }
   }
@@ -200,7 +200,7 @@ issuancerules {
 };)";
       auto report = AttestationCollateral::OpenEnclaveReport();
 
-      auto attestResponse = client.AttestOpenEnclave(report, options);
+      auto attestResponse = client->AttestOpenEnclave(report, options);
       // Because a draft policy was set, the resulting token is unsigned.
       ValidateAttestResponse(
           attestResponse, Azure::Nullable<AttestationData>(), *options.DraftPolicyForAttestation);
@@ -216,7 +216,7 @@ authorizationrules
 issuancerules {
     c:[type=="x-ms-sgx-mrsigner"] => issue(type="custom-name", value=c.value);
 };)";
-      EXPECT_THROW(client.AttestOpenEnclave(report, options), Azure::Core::RequestFailedException);
+      EXPECT_THROW(client->AttestOpenEnclave(report, options), Azure::Core::RequestFailedException);
     }
     else if (type == AttestationType::SgxEnclave)
     {
@@ -234,7 +234,7 @@ issuancerules {
     c:[type=="x-ms-sgx-mrsigner"] => issue(type="custom-name", value=c.value);
 };)";
       auto quote = AttestationCollateral::SgxQuote();
-      auto attestResponse = client.AttestSgxEnclave(quote, options);
+      auto attestResponse = client->AttestSgxEnclave(quote, options);
       ValidateAttestResponse(
           attestResponse, Azure::Nullable<AttestationData>(), *options.DraftPolicyForAttestation);
 
@@ -249,7 +249,7 @@ authorizationrules
 issuancerules {
     c:[type=="x-ms-sgx-mrsigner"] => issue(type="custom-name", value=c.value);
 };)";
-      EXPECT_THROW(client.AttestSgxEnclave(quote, options), Azure::Core::RequestFailedException);
+      EXPECT_THROW(client->AttestSgxEnclave(quote, options), Azure::Core::RequestFailedException);
     }
   }
 
@@ -286,13 +286,13 @@ issuancerules {
                 EXPECT_NE(cert->GetSubjectName().find(m_endpoint), std::string::npos);
               }
             };
-      auto attestResponse = client.AttestOpenEnclave(report, options);
+      auto attestResponse = client->AttestOpenEnclave(report, options);
       ValidateAttestResponse(attestResponse, data);
     }
     else if (type == AttestationType::SgxEnclave)
     {
       auto quote = AttestationCollateral::SgxQuote();
-      auto attestResponse = client.AttestSgxEnclave(quote, {data});
+      auto attestResponse = client->AttestSgxEnclave(quote, {data});
       ValidateAttestResponse(attestResponse, data);
     }
   }